feat : 인증 코드 포커스 이동 기능 구현
