+++ app/src/main/res/layout/fragment_email_verify.xml
@@ -52,30 +52,54 @@
             app:layout_constraintTop_toBottomOf="@+id/label_layout">
 
             <EditText
+                android:id="@+id/signup_code_1"
+                android:inputType="numberDecimal"
+                android:maxLength="1"
+                android:gravity="center"
+                android:textAlignment="center"
                 android:layout_width="0dp"
                 android:layout_height="100dp"
                 android:layout_marginEnd="16dp"
                 android:layout_weight="1"
+                android:textSize="@dimen/large"
                 android:background="@drawable/round_gray_input_bg" />
 
             <EditText
+                android:id="@+id/signup_code_2"
+                android:inputType="numberDecimal"
+                android:maxLength="1"
+                android:gravity="center"
+                android:textAlignment="center"
                 android:layout_width="0dp"
                 android:layout_height="100dp"
                 android:layout_marginEnd="16dp"
                 android:layout_weight="1"
+                android:textSize="@dimen/large"
                 android:background="@drawable/round_gray_input_bg" />
 
             <EditText
+                android:id="@+id/signup_code_3"
+                android:inputType="numberDecimal"
+                android:maxLength="1"
+                android:gravity="center"
+                android:textAlignment="center"
                 android:layout_width="0dp"
                 android:layout_height="100dp"
                 android:layout_marginEnd="16dp"
                 android:layout_weight="1"
+                android:textSize="@dimen/large"
                 android:background="@drawable/round_gray_input_bg" />
 
             <EditText
+                android:id="@+id/signup_code_4"
+                android:inputType="numberDecimal"
+                android:maxLength="1"
+                android:gravity="center"
+                android:textAlignment="center"
                 android:layout_width="0dp"
                 android:layout_height="100dp"
                 android:layout_weight="1"
+                android:textSize="@dimen/large"
                 android:background="@drawable/round_gray_input_bg" />
         </LinearLayout>
